key adapter class
